Fix Issue #17 repeated rescaling of image map.
Repeated scaling of image map will after a while make the image map
coordinates diverge from the location of the airport on the map. To fix this
the coordinates of the image map will be cached and used for calculating
the new image map coordinates when scaling is applie.

This patch also now resizes the airport locator circle so it is not a tiny circle
when zoomed to 200%.