New Landing Page (#28)
* Map stays roughly centered when zooming in and out.

* Scroll compensation on zoom now happening in the same event loop. (#22)

Fix flicker and better centering

- Scroll compensation on zoom now happening in the same event loop (no flickering).
- Removed vi swap files added accidentally.
- Added .gitignore with vi swap files.

---------

Co-authored-by: Sean Eidemiller <sean.eidemiller@oracle.com>

* Add scroll wheel zoom

This patch adds the ability to zoom with the scroll wheel like in BMS.
Also the zoom can now go from 50% to 200% so  greater range with a
default of 100%.

Issue: #5

* Sync dev branch with main (#27)

* Add scroll wheel zoom in and out (#25)

* Map stays roughly centered when zooming in and out.

* Scroll compensation on zoom now happening in the same event loop. (#22)

Fix flicker and better centering

- Scroll compensation on zoom now happening in the same event loop (no flickering).
- Removed vi swap files added accidentally.
- Added .gitignore with vi swap files.

---------

Co-authored-by: Sean Eidemiller <sean.eidemiller@oracle.com>

* Add scroll wheel zoom

This patch adds the ability to zoom with the scroll wheel like in BMS.
Also the zoom can now go from 50% to 200% so  greater range with a
default of 100%.

Issue: #5

---------

Co-authored-by: Sean Eidemiller <sean.eidemiller@oracle.com>
Co-authored-by: MaxWaldorf <tigrou40@hotmail.com>
Co-authored-by: Sean Eidemiller <157770408+rsedevnet@users.noreply.github.com>

* Fix Issue #17 repeated rescaling of image map.

Repeated scaling of image map will after a while make the image map
coordinates diverge from the location of the airport on the map. To fix this
the coordinates of the image map will be cached and used for calculating
the new image map coordinates when scaling is applie.

This patch also now resizes the airport locator circle so it is not a tiny circle
when zoomed to 200%.

---------

Co-authored-by: Sean Eidemiller <sean.eidemiller@oracle.com>
Co-authored-by: MaxWaldorf <tigrou40@hotmail.com>
Co-authored-by: Sean Eidemiller <157770408+rsedevnet@users.noreply.github.com>

* Prepare for 4.38 Maps

New landing page and under construction for 4.38 Maps.

---------

Co-authored-by: Sean Eidemiller <sean.eidemiller@oracle.com>
Co-authored-by: MaxWaldorf <tigrou40@hotmail.com>
Co-authored-by: Sean Eidemiller <157770408+rsedevnet@users.noreply.github.com>