KTO Basic Map Setup
This PR contains an initial prototype for the 4.38 release with the new map_airbases overlay and JS fixes for zooming and scroll zoom. The top-level landing page will line to the KTO maps but the plates are not ready. It also contains the basic KTO framework but uses the 4.37 charts and data in many places.